Extract Forex into separate class
